pyproject.toml
[project]
name = "rethink-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "google-cloud-secret-manager>=2.18.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "uvicorn>=0.24.0",
]
